backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sentence-transformers
numpy
nltk